+++ sender/handlers/__init__.py
@@ -7,6 +7,7 @@ from .image import send_image as send_image
 from .location import send_location as send_location
 from .music import send_music as send_music
 from .plain import send_plain as send_plain
+from .poll import send_poll as send_poll
 from .rps import send_rps as send_rps
 from .shake import send_shake as send_shake
 from .share import send_share as send_share